requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sqlalchemy[asyncio]
aiosqlite
httpx
pydantic
aioapns
passlib[bcrypt]
python-jose[cryptography]
python-multipart
